profile: remove special casing of handle moving
When moving the handle with the mouse, the tried to be smart
about changing the active handle when crossing handles.

To me this always felt weird and it was inconsistence with
mouse move. Simply do nothing special at all. The user
should hopefully get an intiutive grasp of whats going on
when moving on handler across another.

Signed-off-by: Berthold Stoeger <bstoeger@mail.tuwien.ac.at>